[63812] Neural API Support (#15)
This PR brings support for the Nylas Neural API. The Neural API exposes endpoints that perform intelligent analysis on items in your Nylas account. This change will enable users to access the features of the Nylas Neural API and provides some utility functions to help improve the user's workflow. Please note that at this time we are not supporting the Categorize Neural API endpoint for this SDK, but we are looking to add it in the near future.