Fix deserialization errors with "label" field
The label field returned by the neural categorize endpoint is a list of strings (display name) instead of a list of labels as expected with Message-types.

To workaround this we introduce an adapter that will parse the incoming API response and create a new label objects with the display names, then deserialize.

We also added a new constuctor to `Label` to create a shell object with just the display name set.