Fixed Participant status not being sent to the API when set (#68)
The API allows for event participants to have a status set for them on create. In the Java SDK, the setter was deprecated and was not transmitting the value on create. Now we introduced a new Enum with all the possible participant status values (Participant.Status) and a new setter. On creating a new event, the Java SDK will now send the participant status to the API, if set.