Add scheduler support for the EU region (#151)
# Description
This PR enables support for scheduler using the EU region. Depending on
the regional URL set for the Nylas SDK, the region for the scheduler
will get set as well.

# Usage
```java
public class NylasExample {
  public static void main(String[] args) throws Exception {
    // Setup Nylas client, defaulting to US region
    NylasClient client = new NylasClient();
    NylasApplication application = client.account("ACESS_TOKEN");
    Schedulers usScheduler = application.schedulers(); // Scheduler URL pointing to https://api.schedule.nylas.com/

    // Setup Nylas client for the EU region
    NylasClient euClient = new NylasClient.Builder()
        .baseUrl(NylasClient.EU_BASE_URL)
        .build();
    NylasApplication euApplication = euClient.account("ACESS_TOKEN");
    Schedulers euScheduler = euApplication.schedulers(); // Scheduler URL pointing to https://ireland.api.schedule.nylas.com/
  }
}
```

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.