Replace Java with Kotlin + Add Auth, Calendar & Event support (#154)
# Description
This PR shifts the Java SDK into a Kotlin SDK. This is done to enable
friendlier support for Kotlin-powered applications while still keeping
the SDK feeling native to Java, as we're designing this SDK with Java
interoperability in mind. This PR also brings support for:

- Nylas API v3
- Auth methods
- Calendars API CRUD
- Events API CRUD
- Application CRUD

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.