SDK v2 Beta - Bug fixes, cleanup, and in-code documentation (#157)
This PR adds in-code documentation to all our models and methods to help
generate a series of supporting documentation we are calling the "SDK
Reference". The in-code documentation is formatted with KDoc and the
docs are generated using Dokka. This PR also cleans up more files from
the repo deleting un-needed legacy classes and files. Finally, this PR
also fixes the following:
- Refactored Auth methods:
  - Renamed appropriate methods to denote that they use OAuth 2.0
  - Removed support for Auth endpoints that no longer exists
  - Removed support for Providers as they are removed from the API
- Removed the client id/client secret requirement to use the auth
resource
- Make provider optional for building OAuth 2 URL
- Fix token exchange methods
- Fix revoking token
- Fix query params not being set properly
- Add missing serializer for Conferencing
- Add error classes for API errors and SDK errors

[TW-2320](https://nylas.atlassian.net/browse/TW-2320)
[TW-2321](https://nylas.atlassian.net/browse/TW-2321)

<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.

[TW-2320]:
https://nylas.atlassian.net/browse/TW-2320?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TW-2321]:
https://nylas.atlassian.net/browse/TW-2321?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ