Change artifact name and add markdown files (#156)
# Description
This PR has changed the artifact name of the SDK to be consistent with
our other SDKs from `nylas-java-sdk` to just `nylas`. The SDK also now
adds a "beta" tag to indicate the beta state as we are looking to
publish these packages on maven central. The fully qualified name of the
first beta release will now be:

```
com.nylas.sdk:nylas:2.0.0-beta.1
```

We also added a UPGRADE.md file for guiding users on using the new SDK,
as well as updating the README.md.

# JIRA Tickets
[TW-2318](https://nylas.atlassian.net/browse/TW-2318)
[TW-2319](https://nylas.atlassian.net/browse/TW-2319)

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.

[TW-2318]:
https://nylas.atlassian.net/browse/TW-2318?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[TW-2319]:
https://nylas.atlassian.net/browse/TW-2319?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Lrubin <Lrubin@nylas.com>
Co-authored-by: LRubin <sanscontext@users.noreply.github.com>