Final Pre-v2.0.0-beta-1 release fixes (#160)
# Description
This PR sets the Kotlin/Java SDK to being code-complete and ready for
the first beta release. What changed:
- Added availability support
- Added webhook support
- Added missing documentation
- Added missing null initializers for optional params
- Added missing enums
- Cleaned up unused models
- Changed the name `serverUrl` to `apiUri` for more consistency
- Fixed token request required fields

# JIRA Tickets
https://nylas.atlassian.net/browse/AV-1703
https://nylas.atlassian.net/browse/AV-1448

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.