(Kotlin) Rotate secret should be post (#164)
# Description

(Kotlin) Right now rotateSecret is set as PUT method, although it should
POST according to the documentation. This error gets generated when
calling the method.

```
Exception in thread "main" java.lang.IllegalArgumentException: method PUT must have a request body.
```

# License
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.