Fix Exchanging Token always returns 401 (#178)
# Description
The API doesn't like the trailing "charset=utf-8" in the content-type
header. Removing it allows the calls to process properly for auth.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.