Fixed int type being serialized to double sometimes (#177)
# Description
Sometimes when serializing query parameters, it converts an integer to a
double. The decimal produced breaks the API call. We ensure that we are
handling the Double instance specifically.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.