Fix error when sending message or updating draft (#179)
# Description
There were a couple of errors with sending messages and updating drafts.
First was `attachments` was attempting to be serialized when it should
have been marked as Transient. Secondly, we were passing the wrong type
to deserialize the response.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.