v2.0.0 (#202)
# Description
This PR is a culmination of all the PRs that contained v2-related
development, with formal support for the Nylas API v3. This release is
essentially the same as the v2.0.0-beta.5 release found on maven. This
release also formally switches the development language of the SDK from
Java to Kotlin, as well as the artifact name from `nylas-java-sdk` to
`nylas`. Please refer to the readme for a quick start guide on using the
new SDK, as well as links to the upgrade doc as well as the SDK
reference docs.

# Changelog
### BREAKING CHANGES

* Renamed artifact from `nylas-java-sdk` to `nylas`.
* Nylas SDK v2 supports the Nylas API v3 exclusively, dropping support
for any endpoints that are not available in v3.
* Removed all REST calls from models and moved them directly into
resources

### Added

* Full Kotlin support
* Created models for all API resources and endpoints, for all HTTP
methods to reduce confusion on which fields are available for each
endpoint
* Created error classes for the different API errors as well as
SDK-specific errors

### Changed

* Leveraged Moshi annotations for JSON serialization/deserialization as
opposed to manually writing JSON maps
* Removed all REST calls from models and moved them directly into
resources

### Removed

* Non-builder ways for initializing `NylasClient`
* Local Webhook development support is removed due to incompatibility
with the new API version

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.

---------

Co-authored-by: Aaron de Mello <314152+AaronDDM@users.noreply.github.com>
Co-authored-by: Lrubin <Lrubin@nylas.com>
Co-authored-by: LRubin <sanscontext@users.noreply.github.com>
Co-authored-by: Blag <atejada@gmail.com>
Co-authored-by: YIFAN WU <14408339+yifanplanet@users.noreply.github.com>