Change default timeout to match API (90 seconds) (#218)
# Description
The API's default timeout is 90 seconds, so it's best we match that. We
have updated the default value to match this, user can still override
the default.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.