Improved message sending and draft create/update performance (#217)
# Description
This PR improves message send/draft create/update performance by always
defaulting to application/json instead of multipart. Multipart will only
be used for when a request contains a total attachments size of 3mb or
higher.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.