Fixed issue where reminders were not being updated correctly when updating an event (#261)
# Background
There was a typo in the UpdateEventRequest which supplied `override`
instead of `overrides` for the reminders JSON object. However, this data
class was also necessarily duplicated. So I got rid of it and used the
pre-existing data class.

# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.