CUST-3867 changed SendMessageRequest.sendAt to Long (#296)
Added a method overload that converts passed ints to Longs, so that the
change is not breaking. There should be a warning thrown if Kotlin users
pass int literals, but these can be ignored/corrected easily (and should
not be happening anyway). Tested this thoroughly.
# License
<!-- Your PR comment must contain the following line for us to merge the
PR. -->
I confirm that this contribution is made under the terms of the MIT
license and that I have the authority necessary to make this
contribution on behalf of its copyright owner.